move file to the root of the repository
Essentially your repository was placed on the url that github provided
you with. However that means that `index.html` was not at the root of
the page - meaning that your website didnt have the proper sturcture
after being published.

Additionally since by defaault yekyll preprocessor is running  a
'.noyekyll' file has to be created to stop this process

(I also took the liberty of removing some trasfiles from the repository
and adding a .gitignore to discourage their addition in the future)